pyproject.toml
[project]
name = "match-estagios"
version = "0.1.0"
description = ""
authors = [
    {name = "Poc-c-727",email = "pocc727@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "flask-mvc2 (>=0.1.0,<0.2.0)",
    "flask (>=3.1.3,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pymysql (>=1.1.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
